globals: statically initialise the display-mode config (800x600x16)
DAT_00588f48/50/54/60 are the fullscreen display-mode parameters the original
holds as .data constants. The port left them zero-initialised, so
InitDirectDraw's SetDisplayMode(0,0,0,...) fell through to the 32-bit desktop
mode - and SetupZBuffer, which picks a Z-buffer format matching the back
buffer's depth, then found no matching format and failed. DAT_00588f50/54 are
also floats (800.0/600.0, read by the coordinate transforms) but were uint8_t.
Give them their real values (width 800.0, height 600.0, bpp 16, refresh 60,
decoded from orig .data 0x588f48-0x588f60) and pass them to SetDisplayMode
directly instead of the argless FloatToInt64() the recompile turned to garbage.

+++ include/globals.h
@@ -354,9 +354,9 @@ extern uint16_t DAT_00588f40;
 extern uint32_t DAT_00588f44;
 extern uint32_t DAT_00588f48;
 extern uint8_t DAT_00588f4c;
-extern uint8_t DAT_00588f50;
+extern float DAT_00588f50;
 #define _DAT_00588f50 DAT_00588f50
-extern uint8_t DAT_00588f54;
+extern float DAT_00588f54;
 #define _DAT_00588f54 DAT_00588f54
 extern uint32_t DAT_00588f60;
 extern uint8_t DAT_00588f64;
